refactor(editor): Extract assistant code into features
