+++ packages/frontend/editor-ui/src/api/ai.ts
@@ -1,8 +1,12 @@
-import { useAIAssistantHelpers } from '@/composables/useAIAssistantHelpers';
+import { useAIAssistantHelpers } from '@/features/assistant/composables/useAIAssistantHelpers';
 import { AI_ASSISTANT_MAX_CONTENT_LENGTH } from '@/constants';
 import type { ICredentialsResponse } from '@/Interface';
 import type { IRestApiContext } from '@n8n/rest-api-client';
-import type { AskAiRequest, ChatRequest, ReplaceCodeRequest } from '@/types/assistant.types';
+import type {
+	AskAiRequest,
+	ChatRequest,
+	ReplaceCodeRequest,
+} from '@/features/assistant/assistant.types';
 import { makeRestApiRequest, streamRequest } from '@n8n/rest-api-client';
 import { getObjectSizeInKB } from '@/utils/objectUtils';
 import type { IDataObject } from 'n8n-workflow';
