+++ packages/frontend/editor-ui/src/features/assistant/assistant.types.ts
