+++ packages/frontend/editor-ui/src/features/assistant/components/Agent/NodeIssueItem.vue
