+++ packages/frontend/editor-ui/src/features/assistant/components/HubSwitcher.vue
