+++ packages/frontend/editor-ui/src/features/assistant/composables/useAIAssistantHelpers.ts
@@ -7,13 +7,13 @@ import type {
 	NodeOperationError,
 	INode,
 } from 'n8n-workflow';
-import { useWorkflowHelpers } from './useWorkflowHelpers';
+import { useWorkflowHelpers } from '@/composables/useWorkflowHelpers';
 import { useNDVStore } from '@/stores/ndv.store';
 import { useNodeTypesStore } from '@/stores/nodeTypes.store';
 import { executionDataToJson, getMainAuthField, getNodeAuthOptions } from '@/utils/nodeTypesUtils';
-import type { ChatRequest } from '@/types/assistant.types';
+import type { ChatRequest } from '../assistant.types';
 import { useWorkflowsStore } from '@/stores/workflows.store';
-import { useDataSchema } from './useDataSchema';
+import { useDataSchema } from '@/composables/useDataSchema';
 import { AI_ASSISTANT_MAX_CONTENT_LENGTH, VIEWS } from '@/constants';
 import { useI18n } from '@n8n/i18n';
 import type { IWorkflowDb } from '@/Interface';
